+++ CHANGES.rst
@@ -6,6 +6,11 @@ Changes
 
  - Issue #5: Implemented `path.tempdir`, which returns a path object which is
    a temporary directory and context manager for cleaning up the directory.
+ - Issue #12: One can now construct path objects from a list of strings by
+   simply using path.joinpath. For example::
+
+     path.joinpath('a', 'b', 'c') # or
+     path.joinpath(*path_elements)
 
 2.5
 ---